+++ README.md
@@ -1,6 +1,7 @@
 # PSN Hack Club Website
+
 [![Netlify Status](https://api.netlify.com/api/v1/badges/18b91255-f446-4f21-909a-15f78fe0ca8e/deploy-status)](https://app.netlify.com/sites/psnhackclub/deploys)
 
-This project is a website for the PSN Hack Club  
-  
+This the PSN Hack Club's official website.
+
 Visit the website [here](https://psnhackclub.netlify.app/)
