+++ package.json
@@ -1,51 +1,24 @@
 {
   "name": "psn-hack-website",
-  "version": "0.1.0",
   "private": true,
-  "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "prettier": "^2.4.1",
-    "web-vitals": "^1.0.1"
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
+    "next": "12.0.4",
     "node-sass": "4.14.1",
     "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0",
-    "react-progressive-image": "^0.6.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-snap": "^1.23.0",
-    "react-social-icons": "^5.5.1",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "react-social-icons": "^5.6.1",
     "react-spring": "^9.3.2",
     "react-typist": "^2.0.5"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "postbuild": "react-snap"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "devDependencies": {
+    "eslint": "7",
+    "eslint-config-next": "12.0.4"
   }
 }
