--- public/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="PSN Hack Club Official Website"
-    />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <title>PSN Hack Club</title>
-    <link href="https://fonts.cdnfonts.com/css/bahnschrift" rel="stylesheet">
-  </head>
-  <body>
-    <noscript>Welcome to PSN Hack Club's Official Website!</noscript>
-    <div id="root"></div>
-  </body>
-</html>
