--- src/index.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import { hydrate, render } from "react-dom";
-import './index.scss'
-import App from './App'
-
-const rootElement = document.getElementById("root");
-if (rootElement.hasChildNodes()) {
-  hydrate(<App />, rootElement);
-} else {
-  render(<App />, rootElement);
-}
