--- src/pages/About.module.css
@@ -1,55 +1,0 @@
-﻿.container {
-    margin-top: 4rem;
-    height: max-content;
-}
-
-.logo {
-    width: 35%;
-    display: block;
-    border-radius: 1.5rem;
-    box-shadow: rgba(0, 0, 0, 0.2) 0 0 1.5rem;
-    margin: auto auto .5rem;
-}
-
-.iconList {
-    padding: 0;
-    margin-top: 3rem;
-    list-style: none;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    column-gap: 1.5rem;
-}
-
-.icon {
-    border: 2px #fff solid;
-    border-radius: 10000rem; /* Irrationally high value to have it as a circle indefinitely */
-}
-
-@media (min-width: 576px) {
-    .logo {
-        width: 25%;
-        border-radius: 2rem;
-    }
-}
-
-@media (min-width: 768px) {
-    .container {
-        margin-top: 6rem;
-    }
-
-    .logo {
-        border-radius: 2.5rem;
-    }
-}
-
-@media (min-width: 992px) {
-    .container {
-        margin-top: 10rem;
-    }
-
-    .logo {
-        border-radius: 3rem;
-    }
-}
