+++ src/pages/_app.js
@@ -1,0 +1,12 @@
+import '../styles/globals.scss'
+import Layout from "../components/Layout";
+
+function MyApp({Component, pageProps}) {
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  )
+}
+
+export default MyApp
