+++ src/pages/_document.js
@@ -1,0 +1,19 @@
+import Document, {Html, Head, Main, NextScript} from 'next/document'
+
+class MyDocument extends Document {
+  render() {
+    return (
+      <Html>
+        <Head>
+          <link href="https://fonts.cdnfonts.com/css/bahnschrift" rel="stylesheet"/>
+        </Head>
+        <body>
+        <Main/>
+        <NextScript/>
+        </body>
+      </Html>
+    )
+  }
+}
+
+export default MyDocument
