+++ src/styles/Hackathon2.module.scss
@@ -2,17 +2,11 @@
   display: flex;
   flex-grow: 1;
   flex-direction: column;
-}
 
-hr {
-  width: 600px;
-  border: white 1px solid;
-  background-color: white;
-}
-
-p {
-  padding: 0;
-  margin: 0;
+  p {
+    padding: 0;
+    margin: 0;
+  }
 }
 
 .heroContainer {
